+++ src/components/write/wrtieLeft.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import WriteHeader from '@/components/write/writeHeader';
 import WriteToolBar from '@/components/write/writeArticle/writeToolBar';
-import WriteArticle2 from '@/components/write/writeArticle/writeArticle2';
+import WriteArticle from '@/components/write/writeArticle/writeArticle';
 import WriteSendSection from '@/components/write/writeSendSection';
 
 import WrtieStyle from '../../styles/write/write.module.css'
@@ -12,8 +12,7 @@ const wrtieLeft = () => {
             <div className={WrtieStyle.wrtieLeftWrap}>
                 <div className={WrtieStyle.wrtieLeftWrap1}>
                     <WriteHeader />
-                    {/* <WriteToolBar /> */}
-                    <WriteArticle2 />
+                    <WriteArticle />
                 </div>
                 <div className={WrtieStyle.writeSendSection}>
                     <WriteSendSection />
